Store network_config opts to for inventory generation
Some node network_config options defined in the baremetal
deployment yaml definition can't be stored on neutron
resources as tags, because tag strings has a limited max
lenght.

With this change, these options are store in an ansible
inventory like structure in /var/lib/tripleo-network-config.
The data file is re-written on when nodes are provisioned.

tripleo-ansible-inventory will be extended to look for the
file and include the data in the generated inventory.

Partial-Implements: blueprint network-data-v2-ports
Change-Id: Iadf32e4426a763b0a39581eacc7af7d10d65d8fe