Set source IPv4 address from inventory vars
Currently users have to pass in a dictionary of HOSTNAME:SOURCE_IPV4 of
all nodes in the deployment. An example is as follows from THT:

  FrrBgpIpv4LoopbackMap:
    ctrl-1-0: 99.99.1.1
    ctrl-2-0: 99.99.2.1
    ctrl-3-0: 99.99.3.1
    cmp-1-0: 99.99.1.2
    cmp-2-0: 99.99.2.2
    cmp-3-0: 99.99.3.2
    cmp-1-1: 99.99.1.3
    cmp-2-1: 99.99.2.3
    cmp-3-1: 99.99.3.3

This is rather time consuming, prone to typos and requires updating at
node scale up/down. It would be much easier if users could just pass in
the network and have tripleo_frr get the IP from given network.
Snip from tripleo-ansible-inventory.yaml:

  ControllerRack1:
    hosts:
      ctrl-1-0: {ansible_host: 192.168.1.101, canonical_hostname: ctrl-1-0.bgp.ftw,
        main_network_hostname: ctrl-1-0.mainnetwork.bgp.ftw, main_network_ip: 99.99.1.1,
        [...]

Change-Id: I43852cb3570b8cb12a35f4bc641a42ddfd8ad7f1