Allow rhsm subscription test to be overrided in RHOSP deployment.
This allow the master playbook used for update to set
tripleo_redhat_enforce to false on a per role basis on Red Hat
environment.

The default in defaults/main.yml is now "true" so that it keeps its
behavior of being run by default if nothing is changed in the role
definition.

We then avoid running it on other plateform than Red Hat by adding an
explicit test in that tasks/main.yml file.

Overall the behavior is as follows:

  | Red Hat Env | tripleo_enforced     | Test run |
  |-------------+----------------------+----------|
  | True        | Unset                | Yes      |
  | True        | Set to true in role  | Yes      |
  | True        | Set to false in role | No       |
  | False       | Doesn't matter       | No       |

Change-Id: I6268a01d16f8288bf862003d19184fc93b88282a
Partial-Bug: #1912512