Option to remove the self-installed ReaR cron task
By default, the ReaR rpm installs a cronjob on /etc/cron.d/rear with
a default programming. This programming runs a backup job at
1:30AM every day. During this automatic backup, the services are not
appropiately paused. This causes the created backup to have non-valid
data.

As the most recent backup overwrites the older ones, this deletes
the "good" backup taken at operator's discretion with a new one
that cannot be used to restore the control plane state.

This change aims to optionally disable this default rpm behaviour.

Closes-Bug: #1912764
Change-Id: I2582c1ba74ae115a94ecb4524ba34e79ea5b43e8