Deploy RGW, MDS, Ganesha and Ceph Dashboard using cephadm
This change adds the mkspec module which is able to build
a daemon specific spec and apply it against the deployed
Ceph cluster.
In addition, multiple tasks have been added with the purpose
of integrating all the scenario004 services and properly
manage the cephadm playbook flow according to the enabled
services.

Change-Id: I35e57abddc64310a6422174fe191bd328588d7cd