Update legacy network init
While we need to start the legacy network config on boot, we don't need to
try and start it on task invocation. Starting the legacy network on an
already running host will result in failure because ports are un-able to
be rebound. This change removes the start action and just makes sure that
the legacy network service is enabled.

Change-Id: I88cd4fd907262d6a1bedbe7b76bc025eeb45a837
Signed-off-by: Kevin Carter <kecarter@redhat.com>