Don't query neutron ports to determine ssh user
ssh_user already defaults to heat-admin, so there is no reason to query
Neutron to attempt to determine the default ssh user. The default value
can just be used, or be overridden with tripleoclient's
--overcloud-ssh-user.

This removes a hard dependency on Neutron being installed on the
undercloud, which is not actually needed.

Change-Id: Ic56873960a11a7a6d828d9ff588f47fa404b5e16
Signed-off-by: James Slagle <jslagle@redhat.com>