Service virtual IP module and hieradata
Adds a module tripleo_service_vip which manages
a neutron API port resource for service virtuap IPs.
(redis and ovn_dbs) when neutron service is available.

When the service network is 'ctlplane' the module does
a find_port for the controlplane_virtual_ip so that the
ctlplane VIP is used in this case.

When the neutron service is not available the module
looks for a pre-defined ip address in the fixed_ips
option. If present this address is used, if not present
an exception is raised.

The module updates the 'service_vip_vars.yaml' file in
the playbook directory so that include_vars can be used
to load the variables.

The change also updates action/tripleo_all_nodes_data.py
and tripleo_hieradata templates to use the new variable
to source the service virtual IPs for redis and ovn_dbs
services.

Related: blueprint network-data-v2-ports
Depends-On: https://review.opendev.org/777252
Change-Id: I6b2ae7388f8af15f2fd3dcbc5e671c169700fff6