Handle network config override
It was earlier possible to override network config
for undercloud with a custom config using
``net_config_override`` in undercloud.conf or with
``UndercloudNetConfigOverride`` parameter. Though it's
now possible to change ``UndercloudNetworkConfigTemplate``
parameter to override the default config, we probably
still need to support``net_config_override`` and
``UndercloudNetConfigOverride`` for backward compatibility.

Closes-Bug: #1915585
Depends-On: https://review.opendev.org/c/openstack/tripleo-heat-templates/+/775471
Change-Id: Ied2caf9cd5c1b14d30d0badb4e949f620263c96e