Reduce fact gathering
We've seen that large amount of facts for hosts have a direct impact on
task execution as part of the deployment.  This change reduces the
amount of data that we are collecting when we use facts and leverages
more targeted methods to collect the required information.

Change-Id: I49e6ca02c2b4791641fb27ebf258ef6c9d52dd9e
Related-Bug: #1915761
(cherry picked from commit f6c0acc7fe7e4571fc923d85422d7781b6d35151)