Do not run tripleo_ha_wrapper when the HA service has no config_volume
If a service has no config_volume, it means it has not config file, so
there is no point in trying to restart the service when any of the
configuration files change.

Case in point, ovn_dbs currently generates an empty puppet
config_volume. The only reason this works now is that the tripleo plugin
that calls container-puppet.sh, will generate an empty folder even
though no config is generated. Let's make this less fragile.

Change-Id: I5199cf5fd20590a2aa1af5c52ef805067f995366