[tripleo_transfer] Convert testing to use delegated driver
The current driver is podman which creates problems when we
try to use the synchronize module for the file copy between
hosts. To make room for this option, we switch to using the
delegated driver and two inventory hosts (which are localhost)
instead which is more efficient, more portable and opens more
options.

We also add 'any_errors_fatal: true' to the play
arguments to ensure that any error stops the whole test.

The converge playbook is targeted at localhost, rather than
all instances, because that is the way that the role being
tested is used in TripleO.

Related-Bug: #1908425
Related-Bug: rhbz#1904681
Related-Bug: rhbz#1916162

Change-Id: I4d5049ef863c5685b1d817a865a8a44c4429480c
(cherry picked from commit b1dadc218514f2f4798c9f8264498f03a805fd76)