Add cephadm mkspec library and common functions
This patch adds the mkspec library, used by the related ansible
modules when a spec cephadm file is rendered. It includes the
basic fields defined in [1], and will be improved when new
specifications are added.
In addition, the common functions of the existing ceph_{key, pool}
modules are moved into a shared module_utils library that can be
extended as needed.

[1] https://docs.ceph.com/en/latest/mgr/orchestrator/#service-specification

Change-Id: I0a6f9e56c5cac7c46e0e2048bfea546f1dcb7348