Enable fact bits for ceph execution
ceph-ansible still requires that all the facts be collected by default
and inject fact vars is enabled. We want to turn that off for tripleo,
but in the mean time we need to force it back on when in run ansible for
ceph.

Change-Id: I607c29c45148b57dee34741397cf7a16ced8ef78
Related-Bug: #1915761
Related-Bug: #1917621
(cherry picked from commit ff8d309e9560826c918081a337a10524e13f4013)