Fix tripleo_network config conditional
The conditional used in the NetworkConfig block always
evaluates to conclude that network config should be
applied.

Use | int filter on the slurped os-net-config return
code file so that the condition does not always evaluate
to False.

Since the molecule test is'nt working, also change it
to use the delegate driver. Create a dummy0 interface, and
change the IP addresses used to the 203.0.113.0/24
TEST-NET-3[1] range to avoid conflicts.

[1] https://tools.ietf.org/html/rfc5737

Closes-Bug: #1918372
Change-Id: Ia4d663cd298556c8143eab6e799e68c81bc8235e