Configure crush rules according to the detected device classes
This change introduces tasks, modules and the proper flow needed
to manage crush rules (when specified via tht).
The implemented flow reflects the steps done by ceph-ansible when
crush rules are configured.
The modules are imported from ceph-ansible and a just a few changes
are added to make them cephadm/spec compatible.
The ceph_cli is extended in order to allow the crush_rule tasks to
run commands using --admin-daemon and the mon socket.

Change-Id: If51cdfab825c86cd78da4520cb130d93148a1d43