Remove layout overrides and use tripleo-ci templates instead
As part of [1] and in line with all other tripleo repos lets use
the tripleo-ci templates. The files: and irrelevant-files matches
are moved to tripleo-ci in the depends-on. This leaves only the
undercloud-minion job in the overrides as it isn't covered by the
templates, but updates tags.

[1] https://review.opendev.org/q/topic:tripleo-ci-reduce

Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/781322

Change-Id: Id8bdcc314f68fb35050254c6c64fd2317c2555de
(cherry picked from commit e9b7aaa5fc0bae143a04470b09684909672ca072)