Remove layout overrides and use tripleo-ci templates instead
As part of [1] and in line with all other tripleo repos lets use
the tripleo-ci templates. The files: and irrelevant-files matches
are moved to tripleo-ci in the depends-on.

This is not a clean cherrypick as the Train branch does not have
the undercloud-minion job like V/U/Master. It should run for train
[2] so adding here. Also removes the dependency in the gate between
content providers and linting.

[1] https://review.opendev.org/q/topic:tripleo-ci-reduce
[2] https://opendev.org/openstack/tripleo-ci/src/commit/9bf98f65f5d5dbc369c345b9791772897d1bce7b/zuul.d/multinode-jobs.yaml#L208

Depends-On: https://review.opendev.org/c/openstack/tripleo-ci/+/781322

Change-Id: Id8bdcc314f68fb35050254c6c64fd2317c2555de
(cherry picked from commit e9b7aaa5fc0bae143a04470b09684909672ca072)