Add CephHciOsdCount/CephHciOsdType to tripleo_derive_hci_parameters
The tripleo_derive_hci_parameters module uses the devices and
lvm_volumes list in CephAnsibleDisksConfig from Heat to count the
expected number of OSDs. However, with the move to cephadm, the
OSD service spec language is more flexible and the user can pass
a description like "all available block devices". As a result,
the module can no longer count the expected number of OSDs. As
an alternative expose a paramter so the user may directly pass
CephOsdCount. Similarly expose a CephOsdType parameter (HDD, SSD)
since there is no path to lookup the device in Ironic.

Related-Bug: #1920954
Change-Id: Ia6bbdf023e2a0961cd91d3e9f40a8a5a26253ba3