Do not update podman during deployment
Generally speaking we don't expect any package is updated during
deployment. What is worse, we have experienced several issues caused by
change in podman, thus we should be very careful about the timing
when we update it.

This change ensures that the podman package is not updated by
deployment so that it is updated only by proper minor update steps.

Closes-Bug: #1921069
Change-Id: I1f88b982ecb9a94b45e194691d1ce629df3f9aa7