Introduce the network(s) parameter in the ceph spec module
This change introduces a new option on the CephDaemonSpec definition.
Now a list of networks can be specified and rendered using the related
top level field.
The networks field cannot be logically considered part of extra parameter,
because except for the monitors and osds (they have the cluster and public
networks defined), all the other daemons should provide this kind of field.
The ceph_mkspec module and the rgw tasks are modified accordingly and as
per [1] only rgw is supported for now.

[1] https://github.com/ceph/ceph/pull/40048

Change-Id: I71f1a0e995c0d92d73625cd863488fc183bda7b3