Replace include: with include_tasks: in tripleo_container_manager
According to
https://docs.ansible.com/ansible/latest/collections/ansible/builtin/include_module.html
the include directive is supported but they are looking at its removal
("This module will still be supported for some time but we are looking
at deprecating it in the near future.").

Let's switch this to include_tasks as this was the only user inside all
of tripleo-ansible anyways.

Change-Id: If5967a12e31033cc4f7c63565700450de78661df