Create ganesha-export-index before the ceph-nfs daemon is deployed
When ceph_nfs_rados_backend is provided, as per doc [1], for a fresh
setup, we need to make sure to create the Ganesha export index object
as an empty object before starting the Ganesha server.
This change provides the two tasks, already present in ceph-ansible,
to create the index object.
In addition, since the ganesha systemd units are managed by pacemaker,
this change renders the related j2 templates and prepares the ganesha
workdir in the ceph-nfs target hosts.

[1] https://docs.openstack.org/manila/latest/contributor/ganesha.html
[2] https://github.com/ceph/ceph-ansible/blob/master/roles/ceph-nfs/tasks/start_nfs.yml#L15-L31

Change-Id: Ie89c35238d1bb82a28d560d7f0f718f7ee38044f