Backup&Restore Snapshots RFE
The backup&restore process on the controller needs to be done with
the controller stopped. If that condition is not met, we risk
introducing race conditions that could create unrestorable
non-consistent backups. Currently, the B&R process stops the three
controllers to perform a backup, leaving the cloud unmanageable
during the duration of the backup, which can take several minutes.

This change introduces the capability of doing "serial" backups.
Managing Pacemaker lets us take a node out of the controller
Pacemaker cluster, take a backup from it and introduce it again
on the cluster. While that backup is being done, the other two
nodes of the controller cluster continue to be available to
manage the cloud.

Bugzilla: #1877798
Change-Id: I77c4e6accc205cbc962c663babe1018b51d8d266