Add support for collocating Unbound and bind on the same host
In designate deployments, bind and unbound might be collocated and need
to be bound to the same port on the same network. This patch supports
adding a unique IP so that the two services can do that.

Conflicts:
	tripleo_ansible/roles/tripleo_unbound/tasks/main.yml

Change-Id: I445418bb52fa38604c797a9dc69492edb0a41a8c
(cherry picked from commit e8db40f753ed954ba09ed8bf817e282dd57b4caf)