Don't assume Ironic is installed
During "overcloud node provision", pre-provisioned nodes can be
specified with managed:false, in which case the node provision is just a
data transformation process to prepare the inputs for Heat. The ansible
modules should not assume that Ironic is installed in this case, as the
enable_ironic option has existed for many releases already.

Signed-off-by: James Slagle <jslagle@redhat.com>
Depends-On: I3657169a3b519869f7f3594c8c6a6072a058e0c8
Change-Id: I2bc8f3d8090b2753f636832600d3bd51424ba7e7