Add append option to tripleo_derive_hci_parameters module
The module already has the parameter new_heat_environment_path,
which is a path to a file that gets overwritten each time the
module is called. Add append_new_heat_environment_path, which
is a new bool parameter that defaults to false to perserve the
existing behavior. If the new parameter is true, then merge the
new derived paramters into the file without loosing the existing
derived paramters already in the file.

To close bug 1926627, use new_heat_environment_path with the new
option in the existing role loop. Also, remove the unnecessary
task to write out the derived_parameter_result variable because
we now write it in the previous task.

Closes-Bug: #1926627
Change-Id: Iabb0cbda57fabf27d0e5d8b1cc75706785ecd967