Remove "cli" prefix from cli-derive-local-hci-parameters.yml
Playbooks which are prefixed with "cli" are not meant to be
run directly by TripleO users. The python-tripleoclient is
supposed to execute them instead. The playbook being renamed
is not executed by the python-tripleoclient but is instead
meant to be run directly by the user so update the name to
follow the convention.

Old name: cli-derive-local-hci-parameters.yml
New name: derive-local-hci-parameters.yml

Closes-Bug: #1926762
Change-Id: Icea13df0c48028013a7f6e45eabdb5d17cecb1ec