After SFTP backup, the backup file and the iso must be deleted from the machine
Currently, the SFTP backup leaves huge archives on the machine that the
backup has been taken. This change adds a cleanup task to the
backup&restore role. Also, to perform an SFTP backup, it is needed that
the SSH fingerprint of the storage node is downloaded on the node.
Although this could be done by hand, this change automates also this
action.

bz#1955495

Change-Id: I7b8df9cf8a56606f3db54f00af547378e4f4e472