Remove redhat-8.2.yml var file
This will not work when an operator wants to upgrades from osp-16.1 to
osp-16.2:
16.2 is targeting rhel-8.4 with container-tools:3.0, providing a newer
podman version.

During the upgrade process, we end in a state where we're still running
rhel-8.2, with container-tools:3.0 enabled - leading to an issue when it
comes to the podman version we can get.

Removing this file (and backporting this change down to stable/train)
will allow a swift and smooth fix.

As a last word, this "last security" isn't that useful anymore, since we
have a task that ensures we have the correct module stream[1][2].

[1] https://opendev.org/openstack/tripleo-heat-templates/src/branch/master/deployment/undercloud/undercloud-upgrade.yaml#L67
[2] https://opendev.org/openstack/tripleo-heat-templates/src/branch/master/deployment/tripleo-packages/tripleo-packages-baremetal-puppet.yaml#L261

Closes-Bug: #1927242
Change-Id: Id07aa060dd90fdef9a1cd0a44ddbeaee7d561d18