Move the snapshots feature to optional/beta
Previously a commit[1] introduced a new way of doing backup, using
"snapshots", that do not require to bring down the whole controller to
take the backup. On that commit, the old way was removed, which do not
follow the deprecation procedure.

This commits reintroduces the old way of performing backups via stopping
the controllers and move the snapshots feature to an optional,
user-selected feature.

[1] https://review.opendev.org/c/openstack/tripleo-ansible/+/786380

Change-Id: I49944d585825b32052278cd865a74c46dcb5eaa4