Disable facts gathering in net config playbook
Disable the facts gathering in the playbook used
to apply network configuration to provisioned
nodes.

In some cases the fact gatehereing step causes
the playbook to fail in case the node is just
booting and connection is refused for SSH.

Disabling the facts gathering should also speed up
the plabook execution.

Partially-Implements: blueprint network-data-v2-ports
Change-Id: I436c7426bb430307c972445f04a47f8e11bfcc17