Add the networks parameter to the monitoring stack components
As per [1] the ceph_mkspec module accepts the networks parameter
that defines where the specified daemon should be bound.
This patch adds the existing parameter to the monitoring stack
tasks that are supposed to apply node-exporter(s), prometheus
and alertmanager when DashboardEnabled is true.
In addition, due to recent changes in Ceph, both grafana and
the dashboard passwords must be passed via `-i`. This change
makes us able to configure the password via stdin.
Finally, the tls support is added to the grafana component,
which is exposed to the operators.

[1] https://review.opendev.org/783305

Change-Id: I59a74797dc97540b7553a3e74f67e23e6ccc8f6d
(cherry picked from commit 7e7db792d25ef5ee6f8343fda91b42c5ccdd1746)