Add BGP router-id and resolve-via-default options
The BGP Router Identifier is automatically set as the highest IPv4
address configured on the loopback interface or alternatively from a
physical interface. Thus the router ID may be set to an undesired value.
We are better of explicitely, predictably setting it to the main host
IPv4 address.

In addition, allow the nexthop tracking to resolve via the default
route. This is useful when e.g. you want to allow BGP to peer across the
default route, and MP-BGP EVPN use cases.

Change-Id: Ib22700a80bf96e595103e68a51b2bbdf951892bb