Make execution of the backup idempotent
Currently, the playbook stops the services and the pacemaker to perform the backup. If the backup fails for whatever reason, the playbook cannot run again and fails due to the services being down. This patch fixes the issue in two directions. First, by using the block-always ansible feature, we ensure that whenever an error happens, the playbook will restart the pacemaker and the services before exiting.

The second direction is on the execution itself. By controlling certain actions, we ensure that the backup will be succesful not matter what state the services and the pacemaker are when the playbook is ran. This ensures success even in the case an operator stops the ansible execution manually (which will not trigger the 'always' condition).

BZ: #1954818
Change-Id: Id2aff61f219b0c4992f6f0045f1aba2c7d129758
(cherry picked from commit 9a865d176987f82ecafe121f21ba8db9641379d2)