Create allocation if hostname matches display_name
To facilitate migration from nova to nova-less
deployment create an allocation when the display_name
in the ironic nodes instance_info matches the instance
name associated with the baremetal node matches the
hostname provided in the baremtal deployment YAML
definition.

Also add a safeguard to not allow creating allocations
with duplicate names.

Closes-Bug: #1928487
Change-Id: I743201875381e41b41c15569e0b41045651ac9e0