Fix network-v2 node port idempotency
Store the 'name_upper' of the networks in the
network name id map, and use this name when
setting the port name in port definitions.

The port name is checked agains existing ports,
and since THT uses the upper case network name
when setting the port name the same (name_upper)
must be used here.

Closes-Bug: #1928711
Related: blueprint network-data-v2-ports
Change-Id: Iac7f351ce9ee965c17b6bc2956d140efe25ffcf4