Limit cephadm private key distribution to mons/mgrs
It is not necessary to distribute the ceph-admin user's
private SSH key to every host of a Ceph service; only
the hosts running the ceph_mgr and ceph_mon services
need the private key. By default this is limited to
the controller nodes only.

The existing role calls the ceph-admin-user-playbook.yml.
Split it into two calls which use --limit to target the
necessary Ceph service hosts and set distribute_private_key
to true only for mons/mgrs.

Closes-Bug: #1928717
Change-Id: I8343c419c140670f01bdc94b4c8130004bac64e1