Add playbook to perform UC DB backup only.
Currently, the openstack undercloud backup command allows us setting
up rear, nfs and run the Undercloud's backup. However, it might be
useful to just perform a DB backup instead of a full backup of the
Undercloud.

This patch adds a new playbook cli-undercloud-db-backup.yaml which
invokes only the db_backup.yml play from the backup_and_restore
role. This will perform only the MariaDB database backup for the
Undercloud node. As the file name was hardcoded in the play, this
patch also parametrizes it so that it will be possible to run
the command multiple times without override an existing backup.

Change-Id: I62cadfd95c514edf0b982fa652a53feadb24f2b7