Remove become: true from the role
The role is always executed with --become, as it is specified in the
documentation. The CLI also invokes the role with --become parameter. In
this case, we were currently implementing a non consistent become
policy. As almost all the tasks in the role require root access, it is
more consistent to remove all the become: true entries and launch the
role with --become parameter.

Depends-on: Id2aff61f219b0c4992f6f0045f1aba2c7d129758
Change-Id: I263e05c905e3ea6f3f25f4e5a1b8dfb25f6a3f97