Make the monitoring stack ports configurable for each daemon
According to [1] the monitoring stack daemons are able to bind on the
specified {ip}:{port}.
This change turns the previously defined monitoring items into a dict
where the name of the daemon and the port can be specified (the ports
are supposed to come from the tht parameters).
In addition, we need to make sure that the port parameter is evaluated
as an int, otherwise cephadm fails when the spec is applied. For this
reason a new 'normalize_spec' function is introduced, with the purpose
of analyzing the provided payload and make sure the port parameters
are expressed as int. This is required due to the fact that the ansible
option 'ANSIBLE_JINJA2_NATIVE=True' [2], which is supposed to enable
this feature at tripleo-ansible level is missing, but it's a constraint
we wouldn't like to introduce.

[1] https://github.com/ceph/ceph/pull/41444
[2] https://docs.ansible.com/ansible/latest/reference_appendices/config.html#envvar-ANSIBLE_JINJA2_NATIVE

Change-Id: I3f1d0205f2228d59c5e8bc6b33084c64974c2a2b
(cherry picked from commit ccdecf061ebd39e1fb78c5f370613afb0392294e)