Adjust shutdown service management
We currently attempt to manage the triple container related serivces for
startup/shutdown every time we invoke tripleo container manage. This
management should only occur at most once during a deployment. This
change moves the podman drop in fact to where it's used (for systemd
service unit management) and drops the inclusion of the shutdown.yml for
each inclusion of tripleo_container_manage.

Depends-On: https://review.opendev.org/c/openstack/tripleo-heat-templates/+/794334
Change-Id: Ief0729e51997c3eb1dba679cc40ce7bbf4702bc3