Switch from pacemaker standby to pacemaker stop
Currently, the "snapshots" backup is being taken after taking out one
node from the cluster using pcs node standby. The consequence of it is
that the standby property is also being backuped, so when the node is
restored, it appears as standby and it needs to be unstandby manually.

Switching to pcs cluster stop/pcs cluster start does not make the playbook take any
longer, nor it creates an impact on the time the vips take to be
migrated to a different node while provides the advantage that when the
node is restarted (during restore), pacemaker starts correctly without
any other manual step.

Change-Id: Ided482f7eaed8d5e1fac60679ef3ba07bda41c0c