Use serial or parallel execution depending on the kind of backup
Currently, we support two kinds of backups: the "old way" that stops all
the controllers at the same time and then does the backup and the
"snapshots" that only stops one controller at a time.

In the first case, we want to take the backup of all the controllers
in parallel to save time, while in the second case we want to take the
backup of each controller serialized so we do not stop all of them at
the same time.

This change enables this for the openstack overcloud backup CLI action.

Closes-Bug: #1930861
Change-Id: I72b8380025fbd7b22999d4ea938cb9c218f9ad8f